refactor(client): step 2 -- extract _CoreClient plumbing into client/_core.py (#520)
Pure movement. Moves the shared HTTP plumbing verbatim into a new
_CoreClient(BaseHttpClient) base in client/_core.py: __init__, the four
_*_base_url properties, close, __enter__/__exit__, _request,
_get_or_create_sub_client, _queue_request/_query_request/_encrypt_request/
_sync_actions_request and _wait_for_storage_job. KeboolaClient now inherits
_CoreClient; every endpoint-family mixin peeled out in later steps will
inherit the same base so ty statically resolves self._request(...) etc.
with no # type: ignore.

Judgment call: the relocated __enter__ is annotated `-> Self` instead of
the former `-> "KeboolaClient"`. Once __enter__ lives on the _CoreClient
base, `return self` typed as the KeboolaClient subclass is unsound and ty
flags invalid-return-type; `-> Self` is the correct base-class annotation
and matches the existing BaseHttpClient.__enter__ in this repo
(http_base.py already imports typing.Self for exactly this). Behavior is
identical -- the context manager still returns the concrete instance. This
also drops the TYPE_CHECKING import of KeboolaClient that a verbatim
forward ref would have required.

make check green: 4648 passed, 8 skipped. ty clean.
